refactor: remove cycle from Deductions
